# server/swagger.yaml
openapi: '3.0.0'
info:
  title: Events API
  version: 1.0.0
paths:
...
  /banners:
    get:
      summary: Get all banners
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Banner'
        '500':
          description: Internal server error
    post:
      summary: Create a new banner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Banner'
      responses:
        '201':
          description: Banner created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Banner'
        '400':
          description: Bad request
  /banners/show_true:
    get:
      summary: Get banners with show true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Banner'
        '500':
          description: Internal server error
  /banners/show_false:
    get:
      summary: Get banners with show false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Banner'
        '500':
          description: Internal server error
  /banners/{id}:
    get:
      summary: Get a banner by ID
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Banner ID
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Banner'
        '404':
          description: Banner not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a banner by ID
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Banner ID
      responses:
        '200':
          description: Banner deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Banner'
        '404':
          description: Banner not found
        '500':
          description: Internal server error
    put:
      summary: Update a banner by ID
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Banner ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Banner'
      responses:
        '200':
          description: Banner updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Banner'
        '404':
          description: Banner not found
        '400':
          description: Bad request
...
components:
  schemas:
...
    Banner:
      type: object
      properties:
        id_publisher:
          type: number
          required: true
        title:
          type: string
          required: true
        descrizione:
          type: string
          required: true
        budget:
          type: number
          default: 0
        link:
          type: string
          required: true
        clicks:
          type: number
          default: 0
        views:
          type: number
          default: 0
        show:
          type: boolean
          default: true
...